Organize core into package and relocate tests
