+++ src/crypto_analyzer/__init__.py
@@ -1,0 +1,1 @@
+"""crypto_analyzer package"""
