+++ src/crypto_analyzer/core/data.py
