+++ src/crypto_analyzer/core/levels.py
