+++ src/crypto_analyzer/core/signals.py
